+++ src/StackExchange.Redis/RedisFeatures.cs
@@ -77,6 +77,11 @@ namespace StackExchange.Redis
         public bool GetDelete => Version >= v6_2_0;
 
         /// <summary>
+        /// Does GETEX exist?
+        /// </summary>
+        internal bool GetEx => Version >= v6_2_0;
+
+        /// <summary>
         /// Is HSTRLEN available?
         /// </summary>
         public bool HashStringLength => Version >= v3_2_0;
@@ -172,6 +177,11 @@ namespace StackExchange.Redis
         public bool SortedSetPop => Version >= v4_9_1;
 
         /// <summary>
+        /// Is ZRANGESTORE available?
+        /// </summary>
+        public bool SortedSetRangeStore => Version >= v6_2_0;
+
+        /// <summary>
         /// Are Redis Streams available?
         /// </summary>
         public bool Streams => Version >= v4_9_1;
@@ -234,11 +244,6 @@ namespace StackExchange.Redis
         public bool SetPopMultiple => Version >= v3_2_0;
 
         /// <summary>
-        /// The Redis version of the server
-        /// </summary>
-        public Version Version => version ?? v2_0_0;
-
-        /// <summary>
         /// Are the Touch command available?
         /// </summary>
         public bool KeyTouch => Version >= v3_2_1;
@@ -252,6 +257,11 @@ namespace StackExchange.Redis
         /// Do list-push commands support multiple arguments?
         /// </summary>
         public bool PushMultiple => Version >= v4_0_0;
+
+        /// <summary>
+        /// The Redis version of the server
+        /// </summary>
+        public Version Version => version ?? v2_0_0;
 
         /// <summary>
         /// Create a string representation of the available features.